Move spec metrics actually into cert namespace.
Additionally, drop the SpecLoad and SpecFailure metrics; those only had
meaning when certmg would do opportunistic reloads; we no longer do that,
if you want a reload, SIGHUP.  In light of that, those metrics have no
utility.
